# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend vor Rupay",
        "title": "Rupay Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api",
    "paths": {
...
        "/article-transaction": {
            "get": {
                "description": "Retrieve a list of all article transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleTransactions"
                ],
                "summary": "Retrieve all article transactions",
                "responses": {
                    "200": {
                        "description": "List of article transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.ArticleTransaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new article transaction with the provided payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleTransactions"
                ],
                "summary": "Create a new article transaction",
                "parameters": [
                    {
                        "description": "CreateArticleTransaction payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.ArticleTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ArticleTransaction data",
                        "schema": {
                            "$ref": "#/definitions/db.ArticleTransaction"
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/article-transaction/{articleTransactionId}": {
            "get": {
                "description": "Retrieve an article transaction by the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleTransactions"
                ],
                "summary": "Retrieve an article transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article Transaction ID",
                        "name": "articleTransactionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ArticleTransaction data",
                        "schema": {
                            "$ref": "#/definitions/db.ArticleTransaction"
                        }
                    },
                    "404": {
                        "description": "Article transaction not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an article transaction with the provided payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleTransactions"
                ],
                "summary": "Update an article transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article Transaction ID",
                        "name": "articleTransactionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateArticleTransaction payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateArticleTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ArticleTransaction data",
                        "schema": {
                            "$ref": "#/definitions/db.ArticleTransaction"
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Article transaction not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an article transaction by the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ArticleTransactions"
                ],
                "summary": "Delete an article transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article Transaction ID",
                        "name": "articleTransactionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ArticleTransaction deleted successfully"
                    },
                    "404": {
                        "description": "Article transaction not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete ArticleTransaction",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.ArticleTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "article_uuid": {
                    "type": "string"
                },
                "transaction_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "e.ErrorDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field causing the error",
                    "type": "string"
                },
                "issue": {
                    "description": "Description of the issue with the field",
                    "type": "string"
                }
            }
        },
        "e.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string"
                },
                "details": {
                    "description": "List of field-specific errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/e.ErrorDetail"
                    }
                },
                "error": {
                    "description": "Stack Trace",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                }
            }
        },
...
        "schemas.UpdateArticleTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "article_uuid": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "transaction_uuid": {
                    "$ref": "#/definitions/uuid.NullUUID"
                }
            }
        },
...
        "uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if UUID is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}